refactor(kernel): extract pure helpers and tests out of kernel.rs
kernel.rs was a 14130-line file. Most of it is the LibreFangKernel
type and its impl blocks, which legitimately access private state
together. But three sections were self-contained and could be moved
to sibling modules without any visibility changes to the kernel
type itself:

  - workspace_setup (12 free fns, ~430 body lines)
      ensure_workspaces_layout, migrate_legacy_agent_dirs,
      init_git_if_missing, ensure_workspace, safe_path_component,
      has_unsafe_relative_components, resolve_workspace_dir,
      backfill_workspace_dir, generate_identity_files,
      append_daily_memory_log, read_identity_file, gethostname

  - manifest_helpers (9 free fns, ~280 body lines)
      manifest_to_capabilities, apply_thinking_override,
      apply_budget_defaults, default_embedding_model_for_provider,
      infer_provider_from_model, extract_manifest_from_hand_toml,
      toml_enabled_false, shared_memory_agent_id, peer_scoped_key

  - tests (the entire #[cfg(test)] mod tests block, ~1300 lines)

Each extracted file is a sibling module of kernel/mod.rs with the
moved fns promoted from private to pub(super) so the parent module
can still reach them via 'use foo::*'. cron_deliver_response stays
in mod.rs because it borrows &LibreFangKernel and would have
required a pub(super) field exposure to move out.

Net effect:
  kernel.rs: 14130 lines  ->  kernel/mod.rs: 12107 lines (-14.3%)

No behaviour changes. Function bodies were moved verbatim; the only
edits were the visibility keyword on each fn declaration. The kernel
module is now a directory module (kernel/mod.rs) instead of a single
file. tests.rs was reformatted by cargo fmt to remove the now-redundant
extra indentation that was there because the tests block used to be
nested inside 'mod tests { ... }'.
