+++ crates/librefang-kernel/src/kernel/tests.rs
@@ -1,0 +1,1313 @@
+use super::*;
+use futures::stream;
+use librefang_channels::types::{ChannelAdapter, ChannelContent, ChannelType, ChannelUser};
+use librefang_types::approval::{
+    AgentNotificationRule, ApprovalRequest, NotificationConfig, NotificationTarget, RiskLevel,
+};
+use librefang_types::config::DefaultModelConfig;
+use std::collections::HashMap;
+use std::pin::Pin;
+
+struct RecordingChannelAdapter {
+    name: String,
+    channel_type: ChannelType,
+    sent: Arc<std::sync::Mutex<Vec<String>>>,
+}
+
+impl RecordingChannelAdapter {
+    fn new(channel_type: &str) -> Self {
+        Self {
+            name: channel_type.to_string(),
+            channel_type: ChannelType::Custom(channel_type.to_string()),
+            sent: Arc::new(std::sync::Mutex::new(Vec::new())),
+        }
+    }
+}
+
+#[async_trait]
+impl ChannelAdapter for RecordingChannelAdapter {
+    fn name(&self) -> &str {
+        &self.name
+    }
+
+    fn channel_type(&self) -> ChannelType {
+        self.channel_type.clone()
+    }
+
+    async fn start(
+        &self,
+    ) -> Result<
+        Pin<Box<dyn futures::Stream<Item = librefang_channels::types::ChannelMessage> + Send>>,
+        Box<dyn std::error::Error + Send + Sync>,
+    > {
+        Ok(Box::pin(stream::empty()))
+    }
+
+    async fn send(
+        &self,
+        user: &ChannelUser,
+        content: ChannelContent,
+    ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+        if let ChannelContent::Text(text) = content {
+            self.sent
+                .lock()
+                .unwrap()
+                .push(format!("{}:{text}", user.platform_id));
+        }
+        Ok(())
+    }
+
+    async fn stop(&self) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+        Ok(())
+    }
+}
+
+struct EnvVarGuard {
+    key: &'static str,
+}
+
+impl Drop for EnvVarGuard {
+    fn drop(&mut self) {
+        std::env::remove_var(self.key);
+    }
+}
+
+fn set_test_env(key: &'static str, value: &str) -> EnvVarGuard {
+    std::env::set_var(key, value);
+    EnvVarGuard { key }
+}
+
+#[test]
+fn test_collect_rotation_key_specs_dedupes_primary_profile_key() {
+    let _primary = set_test_env("LIBREFANG_TEST_ROTATION_PRIMARY_KEY_A", "key-1");
+    let _secondary = set_test_env("LIBREFANG_TEST_ROTATION_SECONDARY_KEY_A", "key-2");
+    let profiles = [
+        AuthProfile {
+            name: "secondary".to_string(),
+            api_key_env: "LIBREFANG_TEST_ROTATION_SECONDARY_KEY_A".to_string(),
+            priority: 10,
+        },
+        AuthProfile {
+            name: "profile-a".to_string(),
+            api_key_env: "LIBREFANG_TEST_ROTATION_PRIMARY_KEY_A".to_string(),
+            priority: 0,
+        },
+    ];
+
+    let specs = collect_rotation_key_specs(Some(&profiles), Some("key-1"));
+
+    assert_eq!(
+        specs,
+        vec![
+            RotationKeySpec {
+                name: "profile-a".to_string(),
+                api_key: "key-1".to_string(),
+                use_primary_driver: true,
+            },
+            RotationKeySpec {
+                name: "secondary".to_string(),
+                api_key: "key-2".to_string(),
+                use_primary_driver: false,
+            },
+        ]
+    );
+}
+
+#[test]
+fn test_collect_rotation_key_specs_prepends_distinct_primary_and_skips_missing_profiles() {
+    let _secondary = set_test_env("LIBREFANG_TEST_ROTATION_SECONDARY_KEY_B", "key-2");
+    let profiles = [
+        AuthProfile {
+            name: "missing".to_string(),
+            api_key_env: "LIBREFANG_TEST_ROTATION_MISSING_KEY_B".to_string(),
+            priority: 0,
+        },
+        AuthProfile {
+            name: "secondary".to_string(),
+            api_key_env: "LIBREFANG_TEST_ROTATION_SECONDARY_KEY_B".to_string(),
+            priority: 1,
+        },
+    ];
+
+    let specs = collect_rotation_key_specs(Some(&profiles), Some("key-0"));
+
+    assert_eq!(
+        specs,
+        vec![
+            RotationKeySpec {
+                name: "primary".to_string(),
+                api_key: "key-0".to_string(),
+                use_primary_driver: true,
+            },
+            RotationKeySpec {
+                name: "secondary".to_string(),
+                api_key: "key-2".to_string(),
+                use_primary_driver: false,
+            },
+        ]
+    );
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_notify_escalated_approval_prefers_request_route_to() {
+    let dir = tempfile::tempdir().unwrap();
+    let home_dir = dir.path().to_path_buf();
+    std::fs::create_dir_all(home_dir.join("data")).unwrap();
+
+    let explicit_target = NotificationTarget {
+        channel_type: "test".to_string(),
+        recipient: "explicit-recipient".to_string(),
+        thread_id: None,
+    };
+
+    let mut config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+    config.approval.routing = vec![librefang_types::approval::ApprovalRoutingRule {
+        tool_pattern: "shell_*".to_string(),
+        route_to: vec![NotificationTarget {
+            channel_type: "test".to_string(),
+            recipient: "policy-recipient".to_string(),
+            thread_id: None,
+        }],
+    }];
+    config.notification = NotificationConfig {
+        approval_channels: vec![NotificationTarget {
+            channel_type: "test".to_string(),
+            recipient: "global-recipient".to_string(),
+            thread_id: None,
+        }],
+        alert_channels: Vec::new(),
+        agent_rules: vec![AgentNotificationRule {
+            agent_pattern: "*".to_string(),
+            channels: vec![NotificationTarget {
+                channel_type: "test".to_string(),
+                recipient: "agent-rule-recipient".to_string(),
+                thread_id: None,
+            }],
+            events: vec!["approval_requested".to_string()],
+        }],
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+    let adapter = Arc::new(RecordingChannelAdapter::new("test"));
+    let sent = adapter.sent.clone();
+    kernel.channel_adapters.insert("test".to_string(), adapter);
+
+    let req = ApprovalRequest {
+        id: uuid::Uuid::new_v4(),
+        agent_id: "agent-123".to_string(),
+        tool_name: "shell_exec".to_string(),
+        description: "run shell command".to_string(),
+        action_summary: "run shell command".to_string(),
+        risk_level: RiskLevel::High,
+        requested_at: chrono::Utc::now(),
+        timeout_secs: 60,
+        sender_id: None,
+        channel: None,
+        route_to: vec![explicit_target],
+        escalation_count: 1,
+    };
+
+    kernel.notify_escalated_approval(&req, req.id).await;
+
+    let sent = sent.lock().unwrap().clone();
+    assert_eq!(
+        sent.len(),
+        1,
+        "only the explicit request target should be used"
+    );
+    assert!(
+        sent[0].starts_with("explicit-recipient:"),
+        "escalation should use the per-request route_to target"
+    );
+    assert!(
+        !sent[0].contains("policy-recipient")
+            && !sent[0].contains("agent-rule-recipient")
+            && !sent[0].contains("global-recipient")
+    );
+
+    kernel.shutdown();
+}
+
+#[test]
+fn test_manifest_to_capabilities() {
+    let mut manifest = AgentManifest {
+        name: "test".to_string(),
+        description: "test".to_string(),
+        author: "test".to_string(),
+        module: "test".to_string(),
+        ..Default::default()
+    };
+    manifest.capabilities.tools = vec!["file_read".to_string(), "web_fetch".to_string()];
+    manifest.capabilities.agent_spawn = true;
+
+    let caps = manifest_to_capabilities(&manifest);
+    assert!(caps.contains(&Capability::ToolInvoke("file_read".to_string())));
+    assert!(caps.contains(&Capability::AgentSpawn));
+    assert_eq!(caps.len(), 3); // 2 tools + agent_spawn
+}
+
+fn test_manifest(name: &str, description: &str, tags: Vec<String>) -> AgentManifest {
+    AgentManifest {
+        name: name.to_string(),
+        description: description.to_string(),
+        author: "test".to_string(),
+        module: "builtin:chat".to_string(),
+        tags,
+        ..Default::default()
+    }
+}
+
+#[test]
+fn test_send_to_agent_by_name_resolution() {
+    // Test that name resolution works in the registry
+    let registry = AgentRegistry::new();
+    let manifest = test_manifest("coder", "A coder agent", vec!["coding".to_string()]);
+    let agent_id = AgentId::new();
+    let entry = AgentEntry {
+        id: agent_id,
+        name: "coder".to_string(),
+        manifest,
+        state: AgentState::Running,
+        mode: AgentMode::default(),
+        created_at: chrono::Utc::now(),
+        last_active: chrono::Utc::now(),
+        parent: None,
+        children: vec![],
+        session_id: SessionId::new(),
+        tags: vec!["coding".to_string()],
+        identity: Default::default(),
+        onboarding_completed: false,
+        onboarding_completed_at: None,
+        source_toml_path: None,
+        is_hand: false,
+    };
+    registry.register(entry).unwrap();
+
+    // find_by_name should return the agent
+    let found = registry.find_by_name("coder");
+    assert!(found.is_some());
+    assert_eq!(found.unwrap().id, agent_id);
+
+    // UUID lookup should also work
+    let found_by_id = registry.get(agent_id);
+    assert!(found_by_id.is_some());
+}
+
+#[test]
+fn test_find_agents_by_tag() {
+    let registry = AgentRegistry::new();
+
+    let m1 = test_manifest(
+        "coder",
+        "Expert coder",
+        vec!["coding".to_string(), "rust".to_string()],
+    );
+    let e1 = AgentEntry {
+        id: AgentId::new(),
+        name: "coder".to_string(),
+        manifest: m1,
+        state: AgentState::Running,
+        mode: AgentMode::default(),
+        created_at: chrono::Utc::now(),
+        last_active: chrono::Utc::now(),
+        parent: None,
+        children: vec![],
+        session_id: SessionId::new(),
+        tags: vec!["coding".to_string(), "rust".to_string()],
+        identity: Default::default(),
+        onboarding_completed: false,
+        onboarding_completed_at: None,
+        source_toml_path: None,
+        is_hand: false,
+    };
+    registry.register(e1).unwrap();
+
+    let m2 = test_manifest(
+        "auditor",
+        "Security auditor",
+        vec!["security".to_string(), "audit".to_string()],
+    );
+    let e2 = AgentEntry {
+        id: AgentId::new(),
+        name: "auditor".to_string(),
+        manifest: m2,
+        state: AgentState::Running,
+        mode: AgentMode::default(),
+        created_at: chrono::Utc::now(),
+        last_active: chrono::Utc::now(),
+        parent: None,
+        children: vec![],
+        session_id: SessionId::new(),
+        tags: vec!["security".to_string(), "audit".to_string()],
+        identity: Default::default(),
+        onboarding_completed: false,
+        onboarding_completed_at: None,
+        source_toml_path: None,
+        is_hand: false,
+    };
+    registry.register(e2).unwrap();
+
+    // Search by tag — should find only the matching agent
+    let agents = registry.list();
+    let security_agents: Vec<_> = agents
+        .iter()
+        .filter(|a| a.tags.iter().any(|t| t.to_lowercase().contains("security")))
+        .collect();
+    assert_eq!(security_agents.len(), 1);
+    assert_eq!(security_agents[0].name, "auditor");
+
+    // Search by name substring — should find coder
+    let code_agents: Vec<_> = agents
+        .iter()
+        .filter(|a| a.name.to_lowercase().contains("coder"))
+        .collect();
+    assert_eq!(code_agents.len(), 1);
+    assert_eq!(code_agents[0].name, "coder");
+}
+
+#[test]
+fn test_manifest_to_capabilities_with_profile() {
+    use librefang_types::agent::ToolProfile;
+    let manifest = AgentManifest {
+        profile: Some(ToolProfile::Coding),
+        ..Default::default()
+    };
+    let caps = manifest_to_capabilities(&manifest);
+    // Coding profile gives: file_read, file_write, file_list, shell_exec, web_fetch
+    assert!(caps
+        .iter()
+        .any(|c| matches!(c, Capability::ToolInvoke(name) if name == "file_read")));
+    assert!(caps
+        .iter()
+        .any(|c| matches!(c, Capability::ToolInvoke(name) if name == "shell_exec")));
+    assert!(caps.iter().any(|c| matches!(c, Capability::ShellExec(_))));
+    assert!(caps.iter().any(|c| matches!(c, Capability::NetConnect(_))));
+}
+
+#[test]
+fn test_manifest_to_capabilities_profile_overridden_by_explicit_tools() {
+    use librefang_types::agent::ToolProfile;
+    let mut manifest = AgentManifest {
+        profile: Some(ToolProfile::Coding),
+        ..Default::default()
+    };
+    // Set explicit tools — profile should NOT be expanded
+    manifest.capabilities.tools = vec!["file_read".to_string()];
+    let caps = manifest_to_capabilities(&manifest);
+    assert!(caps
+        .iter()
+        .any(|c| matches!(c, Capability::ToolInvoke(name) if name == "file_read")));
+    // Should NOT have shell_exec since explicit tools override profile
+    assert!(!caps
+        .iter()
+        .any(|c| matches!(c, Capability::ToolInvoke(name) if name == "shell_exec")));
+}
+
+#[test]
+fn test_spawn_agent_applies_local_default_model_override() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-local-model-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+    *kernel
+        .default_model_override
+        .write()
+        .expect("default model override lock") = Some(DefaultModelConfig {
+        provider: "ollama".to_string(),
+        model: "Qwen3.5-4B-MLX-4bit".to_string(),
+        api_key_env: String::new(),
+        base_url: Some("http://127.0.0.1:11434/v1".to_string()),
+        ..Default::default()
+    });
+
+    let agent_id = kernel
+        .spawn_agent_inner(
+            AgentManifest {
+                name: "local-model-agent".to_string(),
+                description: "uses local model override".to_string(),
+                author: "test".to_string(),
+                module: "builtin:chat".to_string(),
+                model: ModelConfig {
+                    provider: "default".to_string(),
+                    model: "default".to_string(),
+                    max_tokens: 4096,
+                    temperature: 0.7,
+                    system_prompt: String::new(),
+                    api_key_env: None,
+                    base_url: None,
+                    extra_params: std::collections::HashMap::new(),
+                },
+                ..Default::default()
+            },
+            None,
+            None,
+            None,
+        )
+        .expect("agent should spawn with local model override");
+
+    let entry = kernel.registry.get(agent_id).expect("agent registry entry");
+    // Spawn now stores "default"/"default" so provider changes propagate at
+    // execute time without re-spawning. Concrete resolution happens in
+    // execute_llm_agent, not at spawn.
+    assert_eq!(entry.manifest.model.provider, "default");
+    assert_eq!(entry.manifest.model.model, "default");
+    assert!(entry.manifest.model.base_url.is_none());
+    assert!(entry.manifest.model.api_key_env.is_none());
+
+    kernel.shutdown();
+}
+
+/// Regression: `spawn_agent_inner` must refuse to spawn a child whose
+/// declared capabilities exceed its parent's. Before this check was
+/// pushed down, only `spawn_agent_checked` (tool-runner / WASM host
+/// path) enforced it, and any future caller routing through
+/// `spawn_agent_with_parent` directly (channel handlers, workflow
+/// engines, LLM routing, bulk spawn) would silently bypass the
+/// subset rule and let a restricted parent promote its own
+/// offspring to full privileges.
+#[test]
+fn test_spawn_child_exceeding_parent_is_rejected() {
+    use librefang_types::agent::ManifestCapabilities;
+
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-lineage-reject-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    // Restricted parent: only allowed to invoke `file_read`, no network, no shell.
+    let parent = kernel
+        .spawn_agent_inner(
+            AgentManifest {
+                name: "restricted-parent".to_string(),
+                description: "can only read".to_string(),
+                author: "test".to_string(),
+                module: "builtin:chat".to_string(),
+                capabilities: ManifestCapabilities {
+                    tools: vec!["file_read".to_string()],
+                    ..Default::default()
+                },
+                ..Default::default()
+            },
+            None,
+            None,
+            None,
+        )
+        .expect("parent should spawn as a top-level agent");
+
+    // Malicious child manifest: asks for the wildcard tool +
+    // shell + network — a superset of the parent's single read
+    // capability.
+    let escalation = kernel.spawn_agent_inner(
+        AgentManifest {
+            name: "escalated-child".to_string(),
+            description: "requests full privileges".to_string(),
+            author: "test".to_string(),
+            module: "builtin:chat".to_string(),
+            capabilities: ManifestCapabilities {
+                tools: vec!["*".to_string()],
+                shell: vec!["*".to_string()],
+                network: vec!["*".to_string()],
+                ..Default::default()
+            },
+            ..Default::default()
+        },
+        Some(parent),
+        None,
+        None,
+    );
+    let err = escalation.expect_err("child must be rejected");
+    assert!(
+        format!("{err}").contains("Privilege escalation denied"),
+        "error should mention privilege escalation; got {err}"
+    );
+
+    // Nothing called "escalated-child" should be registered —
+    // the check ran before `register()`.
+    assert!(kernel
+        .registry
+        .list()
+        .iter()
+        .all(|e| e.name != "escalated-child"));
+
+    kernel.shutdown();
+}
+
+/// A child whose capabilities are a strict subset of its parent
+/// still spawns successfully — the check must not refuse legitimate
+/// inheritance. This is the positive counterpart of
+/// `test_spawn_child_exceeding_parent_is_rejected`.
+#[test]
+fn test_spawn_child_with_subset_capabilities_is_allowed() {
+    use librefang_types::agent::ManifestCapabilities;
+
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-lineage-allow-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    let parent = kernel
+        .spawn_agent_inner(
+            AgentManifest {
+                name: "parent-with-file-tools".to_string(),
+                description: "file-reading parent".to_string(),
+                author: "test".to_string(),
+                module: "builtin:chat".to_string(),
+                capabilities: ManifestCapabilities {
+                    tools: vec!["file_read".to_string(), "file_write".to_string()],
+                    ..Default::default()
+                },
+                ..Default::default()
+            },
+            None,
+            None,
+            None,
+        )
+        .expect("parent should spawn");
+
+    let child_id = kernel
+        .spawn_agent_inner(
+            AgentManifest {
+                name: "subset-child".to_string(),
+                description: "narrower read-only child".to_string(),
+                author: "test".to_string(),
+                module: "builtin:chat".to_string(),
+                capabilities: ManifestCapabilities {
+                    tools: vec!["file_read".to_string()],
+                    ..Default::default()
+                },
+                ..Default::default()
+            },
+            Some(parent),
+            None,
+            None,
+        )
+        .expect("subset child should be allowed");
+
+    let entry = kernel.registry.get(child_id).expect("child registered");
+    assert_eq!(entry.parent, Some(parent));
+
+    kernel.shutdown();
+}
+
+/// A child whose `parent` argument points at a registry entry that
+/// doesn't exist must fail closed. This protects against a stale
+/// `AgentId` slipping through (e.g. after a parent is killed mid-
+/// spawn) and silently landing on the non-parent code path.
+#[test]
+fn test_spawn_with_unknown_parent_fails_closed() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-lineage-unknown-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    let ghost_parent = AgentId::new();
+    let result = kernel.spawn_agent_inner(
+        AgentManifest {
+            name: "orphan".to_string(),
+            description: "parent does not exist".to_string(),
+            author: "test".to_string(),
+            module: "builtin:chat".to_string(),
+            ..Default::default()
+        },
+        Some(ghost_parent),
+        None,
+        None,
+    );
+    let err = result.expect_err("unknown parent must fail closed");
+    assert!(
+        format!("{err}").contains("not registered"),
+        "error should indicate parent is not registered; got {err}"
+    );
+
+    kernel.shutdown();
+}
+
+/// Regression: switching an agent's provider via `set_agent_model` must
+/// clear any stale per-agent `api_key_env` / `base_url` overrides. Before
+/// the fix, `update_model_and_provider` only touched `model.provider` and
+/// `model.model`, so an agent that had been booted under a custom default
+/// provider (which seeded those fields onto the manifest) would carry the
+/// old credentials and URL into the new provider, sending requests to the
+/// previous endpoint with the wrong key — surfacing as the upstream's
+/// "Missing Authentication header" 401 (issue #2380).
+#[test]
+fn test_set_agent_model_clears_overrides_when_provider_changes() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-provider-switch-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    // Spawn an agent that already carries the previous provider's
+    // connection overrides — this mirrors the boot-time state of an
+    // agent loaded from disk with provider="default" against a custom
+    // default provider like "cloudverse".
+    let agent_id = kernel
+        .spawn_agent_inner(
+            AgentManifest {
+                name: "switch-provider-agent".to_string(),
+                description: "carries stale overrides from prior provider".to_string(),
+                author: "test".to_string(),
+                module: "builtin:chat".to_string(),
+                model: ModelConfig {
+                    provider: "cloudverse".to_string(),
+                    model: "anthropic-claude-4-5-sonnet".to_string(),
+                    max_tokens: 4096,
+                    temperature: 0.7,
+                    system_prompt: String::new(),
+                    api_key_env: Some("CLOUDVERSE_API_KEY".to_string()),
+                    base_url: Some("https://cloudverse.freshworkscorp.com/api/v1".to_string()),
+                    extra_params: std::collections::HashMap::new(),
+                },
+                ..Default::default()
+            },
+            None,
+            None,
+            None,
+        )
+        .expect("agent should spawn");
+
+    // Sanity: stale overrides are present.
+    let pre = kernel.registry.get(agent_id).expect("agent registry entry");
+    assert_eq!(pre.manifest.model.provider, "cloudverse");
+    assert_eq!(
+        pre.manifest.model.api_key_env.as_deref(),
+        Some("CLOUDVERSE_API_KEY")
+    );
+    assert_eq!(
+        pre.manifest.model.base_url.as_deref(),
+        Some("https://cloudverse.freshworkscorp.com/api/v1")
+    );
+
+    // Switch to an entirely different provider via the same path the
+    // dashboard's model picker uses.
+    kernel
+        .set_agent_model(agent_id, "anthropic/claude-3.5-sonnet", Some("openrouter"))
+        .expect("provider switch should succeed");
+
+    let post = kernel
+        .registry
+        .get(agent_id)
+        .expect("agent registry entry after switch");
+    assert_eq!(post.manifest.model.provider, "openrouter");
+    assert_eq!(
+        post.manifest.model.model, "anthropic/claude-3.5-sonnet",
+        "model name should be updated (and prefix-stripped)"
+    );
+    assert!(
+        post.manifest.model.api_key_env.is_none(),
+        "stale CLOUDVERSE_API_KEY override must be cleared so resolve_driver \
+             falls back to the new provider's key from [provider_api_keys] / convention"
+    );
+    assert!(
+        post.manifest.model.base_url.is_none(),
+        "stale cloudverse base_url override must be cleared so resolve_driver \
+             routes to openrouter's URL from [provider_urls] instead of cloudverse"
+    );
+
+    // Re-applying the same provider (model-only swap) must NOT clear the
+    // override fields — they may be legitimate per-agent overrides on a
+    // single provider.
+    kernel
+        .set_agent_model(agent_id, "anthropic/claude-3.7-sonnet", Some("openrouter"))
+        .expect("same-provider model swap should succeed");
+
+    // Seed an override on the now-openrouter agent so we can confirm the
+    // same-provider branch leaves it alone.
+    kernel
+        .registry
+        .update_model_provider_config(
+            agent_id,
+            "anthropic/claude-3.7-sonnet".to_string(),
+            "openrouter".to_string(),
+            Some("CUSTOM_OPENROUTER_KEY".to_string()),
+            Some("https://my-proxy.example/v1".to_string()),
+        )
+        .expect("seed override");
+
+    kernel
+        .set_agent_model(
+            agent_id,
+            "anthropic/claude-3.7-sonnet-v2",
+            Some("openrouter"),
+        )
+        .expect("same-provider swap should succeed");
+
+    let same_provider = kernel
+        .registry
+        .get(agent_id)
+        .expect("agent after same-provider swap");
+    assert_eq!(
+        same_provider.manifest.model.api_key_env.as_deref(),
+        Some("CUSTOM_OPENROUTER_KEY"),
+        "same-provider swap must preserve per-agent api_key_env override"
+    );
+    assert_eq!(
+        same_provider.manifest.model.base_url.as_deref(),
+        Some("https://my-proxy.example/v1"),
+        "same-provider swap must preserve per-agent base_url override"
+    );
+
+    kernel.shutdown();
+}
+
+#[test]
+fn test_hand_activation_does_not_seed_runtime_tool_filters() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-hand-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+    let instance = match kernel.activate_hand("apitester", HashMap::new()) {
+        Ok(inst) => inst,
+        Err(e) if e.to_string().contains("unsatisfied requirements") => {
+            eprintln!("Skipping test: {e}");
+            kernel.shutdown();
+            return;
+        }
+        Err(e) => panic!("apitester hand should activate: {e}"),
+    };
+    let agent_id = instance.agent_id().expect("apitester hand agent id");
+    let entry = kernel
+        .registry
+        .get(agent_id)
+        .expect("apitester hand agent entry");
+
+    assert!(
+            entry.manifest.tool_allowlist.is_empty(),
+            "hand activation should leave the runtime tool allowlist empty so skill/MCP tools remain visible"
+        );
+    assert!(
+        entry.manifest.tool_blocklist.is_empty(),
+        "hand activation should not set a runtime blocklist by default"
+    );
+
+    kernel.shutdown();
+}
+
+#[test]
+fn test_hand_reactivation_rebuilds_same_runtime_profile() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-reactivation-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    let first_instance = match kernel.activate_hand("apitester", HashMap::new()) {
+        Ok(inst) => inst,
+        Err(e) if e.to_string().contains("unsatisfied requirements") => {
+            eprintln!("Skipping test: {e}");
+            kernel.shutdown();
+            return;
+        }
+        Err(e) => panic!("apitester hand should activate the first time: {e}"),
+    };
+    let first_agent_id = first_instance.agent_id().expect("first apitester agent id");
+    let first_entry = kernel
+        .registry
+        .get(first_agent_id)
+        .expect("first apitester hand agent entry");
+    let first_manifest = first_entry.manifest.clone();
+
+    kernel
+        .deactivate_hand(first_instance.instance_id)
+        .expect("apitester hand should deactivate cleanly");
+
+    let second_instance = match kernel.activate_hand("apitester", HashMap::new()) {
+        Ok(inst) => inst,
+        Err(e) if e.to_string().contains("unsatisfied requirements") => {
+            eprintln!("Skipping test (second activation): {e}");
+            kernel.shutdown();
+            return;
+        }
+        Err(e) => panic!("apitester hand should activate the second time: {e}"),
+    };
+    let second_agent_id = second_instance
+        .agent_id()
+        .expect("second apitester agent id");
+    let second_entry = kernel
+        .registry
+        .get(second_agent_id)
+        .expect("second apitester hand agent entry");
+    let second_manifest = second_entry.manifest.clone();
+
+    assert_eq!(
+        second_manifest.capabilities.tools, first_manifest.capabilities.tools,
+        "reactivation should rebuild the same explicit tool set"
+    );
+    assert_eq!(
+        second_manifest.profile, first_manifest.profile,
+        "reactivation should preserve the same runtime profile"
+    );
+    assert_eq!(
+        second_manifest.tool_allowlist, first_manifest.tool_allowlist,
+        "reactivation should preserve the runtime tool allowlist"
+    );
+    assert_eq!(
+        second_manifest.tool_blocklist, first_manifest.tool_blocklist,
+        "reactivation should preserve the runtime tool blocklist"
+    );
+    assert_eq!(
+        second_manifest.mcp_servers, first_manifest.mcp_servers,
+        "reactivation should preserve MCP server assignments"
+    );
+
+    kernel.shutdown();
+}
+
+#[test]
+fn test_available_tools_returns_empty_when_tools_disabled() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-tools-disabled-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+    let manifest = AgentManifest {
+        name: "no-tools".to_string(),
+        description: "agent with tools disabled".to_string(),
+        author: "test".to_string(),
+        module: "builtin:chat".to_string(),
+        profile: Some(librefang_types::agent::ToolProfile::Full),
+        capabilities: ManifestCapabilities {
+            tools: vec!["file_read".to_string(), "web_fetch".to_string()],
+            ..Default::default()
+        },
+        tools_disabled: true,
+        ..Default::default()
+    };
+
+    let agent_id = kernel.spawn_agent(manifest).expect("spawn should succeed");
+    let tools = kernel.available_tools(agent_id);
+    assert!(
+        tools.is_empty(),
+        "disabled tools should suppress all builtin, skill, and MCP tools"
+    );
+
+    kernel.shutdown();
+}
+
+#[test]
+fn test_available_tools_glob_pattern_matches_mcp_tools() {
+    // Regression: declared tools used exact == match, so "mcp_filesystem_*"
+    // never matched "mcp_filesystem_list_directory" etc. and MCP tools were
+    // silently dropped from available_tools().
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-glob-mcp-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    // Agent with a glob pattern in declared tools — should match builtins
+    let manifest = AgentManifest {
+        name: "glob-tools".to_string(),
+        description: "agent using glob in tools".to_string(),
+        author: "test".to_string(),
+        module: "builtin:chat".to_string(),
+        capabilities: ManifestCapabilities {
+            tools: vec!["file_*".to_string()],
+            ..Default::default()
+        },
+        ..Default::default()
+    };
+
+    let agent_id = kernel.spawn_agent(manifest).expect("spawn should succeed");
+    let tools = kernel.available_tools(agent_id);
+    let names: Vec<&str> = tools.iter().map(|t| t.name.as_str()).collect();
+
+    assert!(
+        names.contains(&"file_read"),
+        "file_* should match file_read, got: {names:?}"
+    );
+    assert!(
+        names.contains(&"file_write"),
+        "file_* should match file_write, got: {names:?}"
+    );
+    assert!(
+        names.contains(&"file_list"),
+        "file_* should match file_list, got: {names:?}"
+    );
+    assert!(
+        !names.contains(&"web_fetch"),
+        "file_* should NOT match web_fetch, got: {names:?}"
+    );
+    assert!(
+        !names.contains(&"shell_exec"),
+        "file_* should NOT match shell_exec, got: {names:?}"
+    );
+
+    kernel.shutdown();
+}
+
+#[test]
+fn test_shell_exec_available_when_declared_in_tools_without_explicit_exec_policy() {
+    // Regression: agents without an explicit exec_policy inherited the global
+    // ExecPolicy whose default mode is Deny, causing shell_exec to be stripped
+    // from available_tools() even when explicitly listed in capabilities.tools.
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-shell-exec-policy-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        // Global exec_policy stays at default (Deny) — this is the scenario
+        // that triggered the bug.
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    let manifest = AgentManifest {
+        name: "shell-agent".to_string(),
+        description: "agent with shell_exec in tools, no exec_policy".to_string(),
+        author: "test".to_string(),
+        module: "builtin:chat".to_string(),
+        capabilities: ManifestCapabilities {
+            tools: vec!["shell_exec".to_string(), "file_read".to_string()],
+            shell: vec!["*".to_string()],
+            ..Default::default()
+        },
+        exec_policy: None, // no explicit policy — must auto-promote
+        ..Default::default()
+    };
+
+    let agent_id = kernel.spawn_agent(manifest).expect("spawn should succeed");
+
+    // Verify exec_policy was promoted to Full
+    let entry = kernel
+        .registry
+        .get(agent_id)
+        .expect("agent must be registered");
+    assert_eq!(
+        entry.manifest.exec_policy.as_ref().map(|p| p.mode),
+        Some(librefang_types::config::ExecSecurityMode::Full),
+        "exec_policy should be auto-promoted to Full when shell_exec is declared"
+    );
+
+    // Verify shell_exec appears in available_tools
+    let tools = kernel.available_tools(agent_id);
+    let names: Vec<&str> = tools.iter().map(|t| t.name.as_str()).collect();
+    assert!(
+        names.contains(&"shell_exec"),
+        "shell_exec must be in available_tools when declared in capabilities.tools, got: {names:?}"
+    );
+
+    kernel.shutdown();
+}
+
+#[test]
+fn test_should_reuse_cached_route_for_brief_follow_up() {
+    assert!(LibreFangKernel::should_reuse_cached_route("fix that"));
+    assert!(LibreFangKernel::should_reuse_cached_route("继续"));
+    assert!(!LibreFangKernel::should_reuse_cached_route("thanks"));
+    assert!(!LibreFangKernel::should_reuse_cached_route(
+        "please write the API design for this service"
+    ));
+}
+
+#[test]
+fn test_assistant_route_key_scopes_sender_and_thread() {
+    let agent_id = AgentId::new();
+    let sender = SenderContext {
+        channel: "telegram".to_string(),
+        user_id: "user-123".to_string(),
+        display_name: "Alice".to_string(),
+        is_group: true,
+        was_mentioned: false,
+        thread_id: Some("thread-9".to_string()),
+        account_id: None,
+        ..Default::default()
+    };
+
+    let with_sender = LibreFangKernel::assistant_route_key(agent_id, Some(&sender));
+    let without_sender = LibreFangKernel::assistant_route_key(agent_id, None);
+
+    assert!(with_sender.contains("telegram"));
+    assert!(with_sender.contains("user-123"));
+    assert!(with_sender.contains("thread-9"));
+    assert_ne!(with_sender, without_sender);
+}
+
+#[test]
+fn test_boot_spawns_assistant_as_default_agent() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home_dir = tmp.path().join("librefang-kernel-default-assistant-test");
+    std::fs::create_dir_all(&home_dir).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+    let agents = kernel.registry.list();
+
+    assert!(
+        agents.iter().any(|entry| entry.name == "assistant"),
+        "fresh kernel boot should auto-spawn an assistant agent"
+    );
+
+    kernel.shutdown();
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_send_message_ephemeral_unknown_agent_returns_not_found() {
+    let dir = tempfile::tempdir().unwrap();
+    let home_dir = dir.path().to_path_buf();
+    std::fs::create_dir_all(home_dir.join("data")).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    // Use a random AgentId that doesn't exist
+    let bogus_id = AgentId::new();
+    let result = kernel.send_message_ephemeral(bogus_id, "hello?").await;
+    assert!(
+        result.is_err(),
+        "ephemeral message to unknown agent should error"
+    );
+
+    kernel.shutdown();
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_send_message_ephemeral_does_not_modify_session() {
+    let dir = tempfile::tempdir().unwrap();
+    let home_dir = dir.path().to_path_buf();
+    std::fs::create_dir_all(home_dir.join("data")).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = LibreFangKernel::boot_with_config(config).expect("Kernel should boot");
+
+    // Find the auto-spawned assistant agent
+    let agents = kernel.registry.list();
+    let assistant = agents
+        .iter()
+        .find(|a| a.name == "assistant")
+        .expect("assistant should exist");
+    let agent_id = assistant.id;
+    let session_id = assistant.session_id;
+
+    // Get session messages before ephemeral call
+    let session_before = kernel.memory.get_session(session_id).unwrap();
+    let msg_count_before = session_before.map(|s| s.messages.len()).unwrap_or(0);
+
+    // Send ephemeral message (will fail because no LLM provider, but that's OK —
+    // the point is the session should remain untouched)
+    let _ = kernel
+        .send_message_ephemeral(agent_id, "what is 2+2?")
+        .await;
+
+    // Check session is unchanged
+    let session_after = kernel.memory.get_session(session_id).unwrap();
+    let msg_count_after = session_after.map(|s| s.messages.len()).unwrap_or(0);
+    assert_eq!(
+        msg_count_before, msg_count_after,
+        "ephemeral /btw message should not modify the real session"
+    );
+
+    kernel.shutdown();
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_spawn_approval_sweep_task_is_idempotent() {
+    let dir = tempfile::tempdir().unwrap();
+    let home_dir = dir.path().to_path_buf();
+    std::fs::create_dir_all(home_dir.join("data")).unwrap();
+
+    let config = KernelConfig {
+        home_dir: home_dir.clone(),
+        data_dir: home_dir.join("data"),
+        ..KernelConfig::default()
+    };
+
+    let kernel = Arc::new(LibreFangKernel::boot_with_config(config).expect("Kernel should boot"));
+
+    Arc::clone(&kernel).spawn_approval_sweep_task();
+    assert!(kernel.approval_sweep_started.load(Ordering::Acquire));
+
+    Arc::clone(&kernel).spawn_approval_sweep_task();
+    assert!(kernel.approval_sweep_started.load(Ordering::Acquire));
+
+    kernel.shutdown();
+    tokio::time::sleep(std::time::Duration::from_millis(25)).await;
+
+    assert!(!kernel.approval_sweep_started.load(Ordering::Acquire));
+}
+
+#[test]
+fn test_evaluate_condition_none() {
+    let tags = vec!["chat".to_string(), "dev".to_string()];
+    assert!(LibreFangKernel::evaluate_condition(&None, &tags));
+}
+
+#[test]
+fn test_evaluate_condition_empty() {
+    let tags = vec!["chat".to_string()];
+    assert!(LibreFangKernel::evaluate_condition(
+        &Some(String::new()),
+        &tags
+    ));
+}
+
+#[test]
+fn test_evaluate_condition_tag_match() {
+    let tags = vec!["chat".to_string(), "dev".to_string()];
+    assert!(LibreFangKernel::evaluate_condition(
+        &Some("agent.tags contains 'chat'".to_string()),
+        &tags,
+    ));
+}
+
+#[test]
+fn test_evaluate_condition_tag_no_match() {
+    let tags = vec!["dev".to_string()];
+    assert!(!LibreFangKernel::evaluate_condition(
+        &Some("agent.tags contains 'chat'".to_string()),
+        &tags,
+    ));
+}
+
+#[test]
+fn test_evaluate_condition_unknown_format() {
+    let tags = vec!["chat".to_string()];
+    // Unknown condition format defaults to false (strict — prevents accidental injection).
+    assert!(!LibreFangKernel::evaluate_condition(
+        &Some("some.unknown.expression".to_string()),
+        &tags,
+    ));
+}
+
+#[test]
+fn test_peer_scoped_key() {
+    // With peer_id: key is namespaced
+    assert_eq!(
+        peer_scoped_key("car", Some("user-123")),
+        "peer:user-123:car"
+    );
+    assert_eq!(
+        peer_scoped_key("prefs.color", Some("u:456")),
+        "peer:u:456:prefs.color"
+    );
+
+    // Without peer_id: key is unchanged
+    assert_eq!(peer_scoped_key("car", None), "car");
+    assert_eq!(peer_scoped_key("global_setting", None), "global_setting");
+}
+
+#[test]
+fn test_apply_thinking_override_none_leaves_manifest_untouched() {
+    let mut manifest = librefang_types::agent::AgentManifest {
+        thinking: Some(librefang_types::config::ThinkingConfig {
+            budget_tokens: 4242,
+            stream_thinking: true,
+        }),
+        ..Default::default()
+    };
+    apply_thinking_override(&mut manifest, None);
+    let cfg = manifest.thinking.as_ref().expect("thinking preserved");
+    assert_eq!(cfg.budget_tokens, 4242);
+    assert!(cfg.stream_thinking);
+}
+
+#[test]
+fn test_apply_thinking_override_force_off_clears_thinking() {
+    let mut manifest = librefang_types::agent::AgentManifest {
+        thinking: Some(librefang_types::config::ThinkingConfig::default()),
+        ..Default::default()
+    };
+    apply_thinking_override(&mut manifest, Some(false));
+    assert!(manifest.thinking.is_none());
+}
+
+#[test]
+fn test_apply_thinking_override_force_on_inserts_default() {
+    let mut manifest = librefang_types::agent::AgentManifest::default();
+    assert!(manifest.thinking.is_none());
+    apply_thinking_override(&mut manifest, Some(true));
+    let cfg = manifest.thinking.as_ref().expect("thinking inserted");
+    assert_eq!(
+        cfg.budget_tokens,
+        librefang_types::config::ThinkingConfig::default().budget_tokens
+    );
+}
+
+#[test]
+fn test_apply_thinking_override_force_on_keeps_existing_budget() {
+    let mut manifest = librefang_types::agent::AgentManifest {
+        thinking: Some(librefang_types::config::ThinkingConfig {
+            budget_tokens: 1234,
+            stream_thinking: false,
+        }),
+        ..Default::default()
+    };
+    apply_thinking_override(&mut manifest, Some(true));
+    let cfg = manifest.thinking.as_ref().expect("thinking preserved");
+    assert_eq!(cfg.budget_tokens, 1234);
+}
